backend/requirements.txt
# Django Framework
Django==4.2.7
djangorestframework==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.0

# CORS handling for React frontend
django-cors-headers==4.3.1

# Image handling
Pillow==10.1.0

# Environment variables
python-decouple==3.8

# Database (PostgreSQL for production)
psycopg2-binary==2.9.9

# Development tools
django-extensions==3.2.3

# API documentation
drf-spectacular==0.26.5

# Testing
coverage==7.3.2

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Static files handling
whitenoise==6.6.0

# Development dependencies (optional - can be in separate dev-requirements.txt)
# flake8==6.1.0
# black==23.10.1
# isort==5.12.0